requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
redis>=5.0.0
prometheus-client>=0.17.0
nvidia-ml-py3>=7.352.0
pytest>=8.2.0,<9
pandas>=2.0.0
numpy>=1.24.0
xgboost>=2.0.0
